Integrate egui rendering into UI shard
# Conflicts:
#	src/gfx/egui/Cargo.toml
#	src/gfx/egui/egui_interop.cpp
