+++ rust/CMakeLists.txt
@@ -5,7 +5,10 @@ if(RUST_SHLISP)
 endif()
 
 # Bindgen dependency
-list(APPEND EXTRA_SOURCES ${SHARDS_DIR}/include/shards.h)
+list(APPEND EXTRA_SOURCES ${SHARDS_DIR}/src/extra/rust_interop.hpp ${SHARDS_DIR}/include/shards.h)
+
+# egui integration sources
+file(GLOB_RECURSE EXTRA_SOURCES "${SHARDS_DIR}/src/gfx/egui/src/*.rs")
 
 if(ANDROID OR LINUX)
   set(NEED_OPENSSL_SYS ON)
