+++ rust/Cargo.toml
@@ -36,6 +36,7 @@ parity-scale-codec = { version = "2.3.1", default-features = false, optional = t
 chacha20poly1305 = { version = "0.9.0", optional = true }
 egui = { version = "0.18.1", optional = true }
 half = { version = "1.8.2" }
+egui-gfx = { path = "../src/gfx/egui"}
 
 [target.'cfg(not(target_arch="wasm32"))'.dependencies]
 dlopen = { version = "0.1.8", optional = true }
