+++ src/extra/gui.hpp
@@ -1,0 +1,10 @@
+#ifndef CF60BD3C_24BE_4A26_A4D8_244E82E6A1DE
+#define CF60BD3C_24BE_4A26_A4D8_244E82E6A1DE
+
+namespace shards {
+namespace Gui {
+
+}
+} // namespace shards
+
+#endif /* CF60BD3C_24BE_4A26_A4D8_244E82E6A1DE */
