+++ src/gfx/egui/egui_interop.cpp
@@ -1,4 +1,5 @@
 #include "egui_interop.hpp"
+#include "egui_render_pass.hpp"
 #include "gfx/params.hpp"
 #include <gfx/enums.hpp>
 #include <gfx/pipeline_step.hpp>
@@ -22,7 +23,6 @@
 using namespace gfx;
 
 namespace egui {
-
 std::vector<MeshVertexAttribute> Vertex::getAttributes() {
   return std::vector<MeshVertexAttribute>{
       MeshVertexAttribute("position", 2),
@@ -133,88 +133,37 @@ struct TextureManager {
   void free(TextureId id) { textures.erase(id.id); }
 };
 
-struct EguiRendererImpl {
-  Renderer &renderer;
+struct RendererImpl {
   MeshPool meshPool;
   TextureManager textures;
+  std::vector<TextureId> pendingTextureFrees;
 
   MeshFormat meshFormat;
-  std::shared_ptr<DrawQueue> queue;
-  ViewPtr view;
-  FeaturePtr uiFeature;
-  PipelineSteps pipelineSteps;
 
-  EguiRendererImpl(Renderer &renderer) : renderer(renderer) {
-    uiFeature = createUIFeature();
-    queue = std::make_shared<DrawQueue>();
-    view = std::make_shared<View>();
-
+  RendererImpl() {
     meshFormat = {
         .primitiveType = PrimitiveType::TriangleList,
         .windingOrder = WindingOrder::CCW,
         .indexFormat = IndexFormat::UInt32,
         .vertexAttributes = Vertex::getAttributes(),
     };
-
-    auto drawableStep = makeDrawablePipelineStep(RenderDrawablesStep{
-        .drawQueue = queue,
-        .features = std::vector<FeaturePtr>{uiFeature},
-        .sortMode = SortMode::Queue,
-    });
-    pipelineSteps = PipelineSteps{drawableStep};
   }
 
-  static FeaturePtr createUIFeature() {
-    auto uiFeature = std::make_shared<Feature>();
-    auto code = shader::blocks::makeCompoundBlock();
+  void addPendingTextureFree(TextureId id) { pendingTextureFrees.push_back(id); }
 
-    code->append(shader::blocks::Header(R"(
-fn linear_from_srgb(srgb: vec3<f32>) -> vec3<f32> {
-  let srgb_bytes = srgb * 255.0;
-  let cutoff = srgb_bytes < vec3<f32>(10.31475);
-  let lower = srgb_bytes / vec3<f32>(3294.6);
-  let higher = pow((srgb_bytes + vec3<f32>(14.025)) / vec3<f32>(269.025), vec3<f32>(2.4));
-  return select(higher, lower, cutoff);
-})"));
-    code->appendLine("var vp = ", shader::blocks::ReadBuffer("viewport", shader::FieldTypes::Float4, "view"));
-    code->appendLine("var p1 = ", shader::blocks::ReadInput("position"), " * (1.0 / vp.zw)");
-    code->appendLine("p1.y = -p1.y;");
-    code->appendLine("p1 = p1 * 2.0 + vec2<f32>(-1.0, 1.0)");
-    code->appendLine("var p4 = vec4<f32>(p1.xy, 0.0, 1.0)");
-    code->appendLine(shader::blocks::WriteOutput("position", shader::FieldTypes::Float4, "p4"));
-    code->appendLine("var color = ", shader::blocks::ReadInput("color"));
-    code->appendLine(
-        shader::blocks::WriteOutput("color", shader::FieldTypes::Float4, "vec4<f32>(linear_from_srgb(color.xyz), color.a)"));
-
-    uiFeature->shaderEntryPoints.emplace_back("baseTransform", ProgrammableGraphicsStage::Vertex, std::move(code));
-
-    shader::FieldType flagsFieldType = shader::FieldType(ShaderFieldBaseType::UInt32, 1);
-    code = shader::blocks::makeCompoundBlock();
-    code->appendLine("var color = ", shader::blocks::ReadInput("color"));
-    code->appendLine("var texColor = ", shader::blocks::SampleTexture("color"));
-    code->appendLine("var isFont = ", shader::blocks::ReadBuffer("flags", flagsFieldType), " & 1u");
-    code->append("if(isFont != 0u) { texColor = vec4<f32>(texColor.xxx, texColor.x); }\n");
-    code->appendLine(shader::blocks::WriteOutput("color", shader::FieldTypes::Float4, "color * texColor"));
-    uiFeature->shaderEntryPoints.emplace_back("color", ProgrammableGraphicsStage::Fragment, std::move(code));
-
-    uiFeature->textureParams.emplace_back("color");
-    uiFeature->shaderParams.emplace_back("flags", flagsFieldType, uint8_t(0));
-
-    uiFeature->state.set_depthWrite(false);
-    uiFeature->state.set_depthCompare(WGPUCompareFunction_Always);
-    uiFeature->state.set_culling(false);
-    uiFeature->state.set_blend(BlendState{
-        .color = BlendComponent::AlphaPremultiplied,
-        .alpha = BlendComponent::Opaque,
-    });
-    return uiFeature;
+  void processPendingTextureFrees() {
+    for (auto &id : pendingTextureFrees) {
+      SPDLOG_INFO("textureFree {}", (uint64_t)id.id);
+      textures.free(id);
+    }
+    pendingTextureFrees.clear();
   }
 };
 
-EguiRenderer::EguiRenderer(gfx::Renderer &renderer) { impl = std::make_shared<EguiRendererImpl>(renderer); }
-void EguiRenderer::render(const FullOutput &output) {
+Renderer::Renderer() { impl = std::make_shared<RendererImpl>(); }
+void Renderer::render(const FullOutput &output, const gfx::DrawQueuePtr &drawQueue) {
   impl->meshPool.reset();
-  impl->queue->clear();
+  impl->processPendingTextureFrees();
 
   // Update textures before render
   auto &textureUpdates = output.textureUpdates;
@@ -240,34 +189,35 @@ void EguiRenderer::render(const FullOutput &output) {
         drawable->parameters.set("flags", uint32_t(0x1));
       }
     }
-    impl->queue->add(drawable);
+    drawQueue->add(drawable);
   }
 
-  impl->renderer.render(impl->view, impl->pipelineSteps);
-
-  // Free textures after render
+  // Store texture id's to free
+  // these are executed at the start the next time render() is called
   for (size_t i = 0; i < textureUpdates.numFrees; i++) {
     auto &id = textureUpdates.frees[i];
-    SPDLOG_INFO("textureFree {}", (uint64_t)id.id);
-    impl->textures.free(id);
+    impl->addPendingTextureFree(id);
   }
 }
 
-void EguiRenderer::getScreenRect(Rect& outScreenRect) {
-  Context &context = impl->renderer.getContext();
+Renderer *Renderer::create() { return new Renderer(); }
+void Renderer::destroy(Renderer *renderer) { delete renderer; }
+} // namespace egui
+
+namespace gfx {
+Context &Renderer_getContext(Renderer &renderer) { return renderer.getContext(); }
+egui::Rect Context_getScreenRect(gfx::Context &context) {
   Window &window = context.getWindow();
   float2 drawScale = window.getDrawScale();
   float2 screenSize = window.getDrawableSize() / std::max(drawScale.x, drawScale.y);
-  outScreenRect = Rect{
-      .min = Pos2{0.0f, 0.0f},
-      .max = Pos2{screenSize.x, screenSize.y},
+  return egui::Rect{
+      .min = egui::Pos2{0.0f, 0.0f},
+      .max = egui::Pos2{screenSize.x, screenSize.y},
   };
 }
-
-float EguiRenderer::getScale() {
-  Context &context = impl->renderer.getContext();
+float Context_getDrawScale(gfx::Context &context) {
   Window &window = context.getWindow();
   float2 drawScale = window.getDrawScale();
   return std::max(drawScale.x, drawScale.y);
 }
-} // namespace egui
+} // namespace gfx
